requirements.txt
aiofiles==0.5.0
alembic==1.4.2
amqp==2.5.2
anytree==2.8.0
apipkg==1.5
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
billiard==3.6.3.0
blessings==1.7
bpython==0.19
cachetools==4.1.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.1
cryptography==2.9.2
cssselect==1.1.0
cssutils==1.0.2
curtsies==0.3.1
decorator==4.4.2
distlib==0.3.0
dnspython==1.16.0
docopt==0.6.2
email-validator==1.1.1
emails==0.5.15
entrypoints==0.3
envparse==0.2.0
execnet==1.7.1
fancycompleter==0.9.1
fastapi==0.54.2
flake8==3.8.2
future-fstrings==1.2.0
github3.py==1.3.0
glob2==0.7
greenlet==0.4.15
h11==0.9.0
httptools==0.1.1
idna==2.9
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
johnnydep==1.5
jwcrypto==0.7
kombu==4.6.8
lazy-object-proxy==1.4.3
lxml==4.5.1
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.3.0
networkx==2.4
oyaml==0.9
packaging==20.4
parse==1.15.0
parse-type==0.5.2
parso==0.7.0
passlib==1.7.2
pathtools==0.1.2
pdbpp==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.5.0.1
pluggy==0.13.1
premailer==3.7.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pwnedpass==0.0.2
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.5.1
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pyrepl==0.9.0
pytest==5.4.2
pytest-bdd==3.3.0
pytest-clarity==0.3.0a0
pytest-console-scripts==0.2.0
pytest-cov==2.9.0
pytest-dependency==0.5.1
pytest-depends==1.0.1
pytest-dotenv==0.4.0
pytest-instafail==0.4.1.post0
pytest-lazy-fixture==0.6.3
pytest-picked==0.4.4
pytest-rerunfailures==9.0
pytest-runner==5.2
pytest-watch==4.2.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
sentry-asgi==0.2.0
sentry-sdk==0.14.4
six==1.15.0
SQLAlchemy==1.3.17
sqlalchemy-mixins==1.2.1
starlette==0.13.2
structlog==20.1.0
tabulate==0.8.7
termcolor==1.1.0
toml==0.10.1
traitlets==4.3.3
transaction==3.0.0
uritemplate==3.0.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
vine==1.3.0
watchdog==0.10.2
wcwidth==0.1.9
websockets==8.1
wimpy==0.6
wmctrl==0.3
wrapt==1.12.1
zope.interface==5.1.0
